release: v1.14.0 — the OpenSNES DX release (#204)
Issues #168-#181 via PRs #190-#203: full CLI/MCP parity, symbols v2 (SPC space), breakpoints v2, search sessions, pokes + per-frame freezes, call_stack(), and the luna test manifest runner. Locally validated: full gate + 90-test golden suite + smoke goldens. Closes #168. Closes #169. Closes #170. Closes #171. Closes #172. Closes #173. Closes #174. Closes #175. Closes #176. Closes #177. Closes #178. Closes #179. Closes #180. Closes #181.

+++ crates/luna-cli/src/parsers.rs
@@ -260,7 +260,7 @@ const FRAME_STEP_BUDGET: u64 = 200_000;
 ///
 /// Returns `None` when the spec is not APU-prefixed, so the caller can
 /// fall through to the CPU-bus and symbol forms.
-pub(crate) fn parse_apu_peek_spec(spec: &str) -> Option<Result<(u16, u16), String>> {
+pub(crate) fn parse_apu_peek_spec(spec: &str) -> Option<Result<(u16, u32), String>> {
     let rest = spec
         .trim()
         .strip_prefix("APU:")
@@ -274,7 +274,7 @@ pub(crate) fn parse_apu_peek_spec(spec: &str) -> Option<Result<(u16, u16), Strin
         Ok(v) => v,
         Err(e) => return Some(Err(format!("bad ARAM offset `{off_s}`: {e}"))),
     };
-    let count = match u16::from_str_radix(count_s.trim(), 16) {
+    let count = match u32::from_str_radix(count_s.trim(), 16) {
         Ok(v) => v,
         Err(e) => return Some(Err(format!("bad count `{count_s}`: {e}"))),
     };
